+++ tests/test_parser.py
@@ -18,6 +18,17 @@ class TestParser(unittest.TestCase):
         
         pass
 
+    def test_parserraw_zi_lines(self):
+
+        case = os.path.join('data', 'psse_sample_case.raw')
+
+        if os.path.isfile(case):
+            net = pf.ParserRAW().parse(case)
+            bus1 = net.get_bus_from_number(3006)
+            bus2 = net.get_bus_from_number(153)
+        else:
+            raise unittest.skipTest('no .m files')
+        
     def test_parserraw_write(self):
 
         tested = False
@@ -47,7 +58,7 @@ class TestParser(unittest.TestCase):
             tested = True
             
         if not tested:
-            self.skipTest("no .raw files")
+            raise unittest.skipTest("no .raw files")
 
     def test_case118_m(self):
 
@@ -59,7 +70,7 @@ class TestParser(unittest.TestCase):
             net_m = pf.Parser(case_m).parse(case_m, num_periods=2)
             pf.tests.utils.compare_networks(self, net_mat, net_m)
         else:
-            self.skipTest('no .m files')
+            raise unittest.skipTest('no .m files')
 
     def test_pyparsermat_write(self):
 
@@ -86,7 +97,7 @@ class TestParser(unittest.TestCase):
             pf.tests.utils.compare_networks(self, net1, net2)
             tested = True
         if not tested:
-            self.skipTest("no .m files")
+            raise unittest.skipTest("no .m files")
 
     def test_ieee25_raw(self):